pyproject.toml
[project]
name = "crypto-trading-bot"
version = "0.1.0"
description = "Premium Crypto Trading Bot using Smart Money Concepts"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "python-telegram-bot",
    "aiohttp",
    "ta",
]